+++ packages/core/src/plugins/day-grid/options.js
@@ -1,0 +1,13 @@
+import {assign} from '#lib';
+
+/**
+ * DayGrid + ResourceTimeline
+ */
+export function createDROptions(options) {
+    if (!('weekNumbers' in options)) {
+        assign(options, {
+            weekNumbers: false,
+            weekNumberContent: undefined
+        });
+    }
+}